requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# === JIM THE MENTOR - AI Mentorship Dependencies ===
anthropic>=0.25.0
openai>=1.0.0
elevenlabs>=0.2.26
replicate>=0.15.0
requests>=2.31.0
pillow>=10.0.0

# === AMBASSADOR PROGRAM DEPENDENCIES ===
supabase>=2.0.0
PyJWT>=2.8.0
